requirements.txt
google.cloud
firebase-admin
